api/requirements.txt
aiofiles==23.1.0
aniso8601==9.0.1
blinker==1.5
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
priority==2.0.0
pyodbc==4.0.39
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
toml==0.10.2
typing_extensions==4.5.0
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.15.0
gunicorn==20.1.0